fix(myparcel): dont import anything from main  plugin file
