+++ packages/vendure-plugin-myparcel/src/myparcel.plugin.ts
@@ -1,49 +1,18 @@
-import { OrderLine, PluginCommonModule, VendurePlugin } from '@vendure/core';
-import { myparcelHandler } from './api/myparcel.handler';
-import { MyparcelController } from './api/myparcel.controller';
-import path from 'path';
+import { PluginCommonModule, VendurePlugin } from '@vendure/core';
 import { AdminUiExtension } from '@vendure/ui-devkit/compiler';
+import path from 'path';
 import { MyparcelConfigEntity } from './api/myparcel-config.entity';
-import { MyparcelService } from './api/myparcel.service';
-import { PLUGIN_INIT_OPTIONS } from './constants';
 import {
-  adminSchema,
   MyparcelAdminResolver,
+  adminSchema,
   myparcelPermission,
 } from './api/myparcel.admin.graphql';
+import { MyparcelController } from './api/myparcel.controller';
+import { myparcelHandler } from './api/myparcel.handler';
+import { MyparcelService } from './api/myparcel.service';
 import { MyParcelShopResolver, shopSchema } from './api/myparcel.shop.graphql';
-
-export interface MyparcelConfig {
-  vendureHost: string;
-  /**
-   * Update webhook in MyParcel platform on Vendure startup or not
-   */
-  syncWebhookOnStartup?: boolean;
-  /**
-   * If you ship outside the EU, you should implement this function to get
-   * {@link CustomsInformation} per order item
-   */
-  getCustomsInformationFn?: (orderLine: OrderLine) => CustomsInformation;
-}
-
-/**
- * Customs information per item for overseas shipping
- * https://myparcelnl.github.io/api/#7_E
- */
-export interface CustomsInformation {
-  /**
-   * Weight in grams per unit
-   */
-  weightInGrams: number;
-  /**
-   * International Standard Industry Classification
-   */
-  classification: string;
-  /**
-   * Country of origin. Has to be a capitalized country code, I.E. 'NL'
-   */
-  countryCodeOfOrigin: string;
-}
+import { MyparcelConfig } from './api/types';
+import { PLUGIN_INIT_OPTIONS } from './constants';
 
 @VendurePlugin({
   imports: [PluginCommonModule],
