+++ packages/vendure-plugin-myparcel/vitest.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: './test/e2e.spec.ts',
+    include: ['./test/e2e.spec.ts'],
   },
   plugins: [
     // SWC required to support decorators used in test plugins
